+++ R/mult-vec.R
@@ -35,10 +35,14 @@ big_prodVec <- function(X, y.col,
                         ind.row = rows_along(X),
                         ind.col = cols_along(X),
                         center = NULL,
-                        scale = NULL) {
+                        scale = NULL,
+                        ncores = 1) {
 
   check_args()
   assert_lengths(y.col, ind.col)
+
+  if (length(ind.row) == 0 || length(ind.col) == 0)
+    return(rep(0, length(ind.row)))
 
   if (!is.null(scale)) {
     assert_lengths(scale, ind.col)
@@ -49,7 +53,10 @@ big_prodVec <- function(X, y.col,
     center2 <- drop(crossprod(as_vec(center), y.col))
   }
 
-  res <- pMatVec4(X, y.col, ind.row, ind.col)
+  res <- big_parallelize(X, function(X, ind, y.col, ind.row, ind.col) {
+    pMatVec4(X, y.col[ind], ind.row, ind.col[ind])
+  }, p.combine = plus, ind = seq_along(ind.col), ncores = ncores,
+  y.col = y.col, ind.row = ind.row, ind.col = ind.col)
 
   `if`(is.null(center), res, res - center2)
 }
@@ -91,10 +98,14 @@ big_cprodVec <- function(X, y.row,
                          ind.row = rows_along(X),
                          ind.col = cols_along(X),
                          center = NULL,
-                         scale = NULL) {
+                         scale = NULL,
+                         ncores = 1) {
 
   check_args()
   assert_lengths(y.row, ind.row)
+
+  if (length(ind.row) == 0 || length(ind.col) == 0)
+    return(rep(0, length(ind.col)))
 
   if (!is.null(scale)) {
     assert_lengths(scale, ind.col)
@@ -105,7 +116,10 @@ big_cprodVec <- function(X, y.row,
     center2 <- sum(y.row) * as_vec(center)
   }
 
-  res <- cpMatVec4(X, y.row, ind.row, ind.col)
+  res <- big_parallelize(X, function(X, ind, y.row, ind.row) {
+    cpMatVec4(X, y.row, ind.row, ind)
+  }, p.combine = 'c', ind = ind.col, ncores = ncores,
+  y.row = y.row, ind.row = ind.row)
 
   if (!is.null(center)) res <- res - center2
   if (!is.null(scale))  res <- res / scale