+++ R/summary.R
@@ -25,6 +25,8 @@ summary.big_sp_list <- function(object, best.only = FALSE, sort = FALSE, ...) {
   res <- foreach(mods = object, .combine = "rbind") %do% {
     tibble::tibble(
       alpha = mods[[1]]$alpha,
+      power_adaptive = mods[[1]]$power_adaptive,
+      power_scale = mods[[1]]$power_scale,
       validation_loss = mean(sapply(mods, function(mod) min(mod$loss.val))),
       intercept = mean(sapply(mods, function(mod) mod$intercept)),
       beta = list(rowMeans(sapply(mods, function(mod) mod$beta))),
@@ -34,10 +36,12 @@ summary.big_sp_list <- function(object, best.only = FALSE, sort = FALSE, ...) {
     )
   }
 
+  order_res <- order(signif(res$validation_loss, 4), res$nb_var)
+
   if (best.only) {
-    res[which.min(res$validation_loss), ]
+    res[order_res[1], ]
   } else if (sort) {
-    res[order(res$validation_loss, -res$alpha), ]
+    res[order_res, ]
   } else {
     res
   }